nl-report-generator/requirements.txt
google-cloud-aiplatform
google-cloud-storage
psycopg2-binary
sqlalchemy
pandas
openpyxl
XlsxWriter
python-pptx
reportlab
flask
streamlit
requests
Faker